refactor `executeActionCallbackHandler`
